API update
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@55657 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/session/session.c
@@ -529,7 +529,7 @@ static void php_session_save_current_state(TSRMLS_D)
 	int vallen;
 	int ret = FAILURE;
 	char *variable;
-	ulong variable_len;
+	uint variable_len;
 	ulong num_key;
 	HashPosition pos;
